Add release prompts file
